portfolio_manager/requirements.txt
aerich==0.7.1
aiosqlite==0.17.0
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==3.4.8
dictdiffer==0.9.0
ecdsa==0.18.0
fastapi==0.92.0
finnhub-python==2.4.16
h11==0.14.0
idna==3.4
iso8601==1.1.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
PyJWT==1.7.1
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tomlkit==0.11.6
tortoise-orm==0.19.3
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
